BANYA UPDATE LAGI YUNUS HEHEHE

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\DashboardController;
+use App\Http\Controllers\FakultasController;
 use App\Http\Controllers\KHSController;
 use App\Http\Controllers\KRSController;
 use App\Http\Controllers\ProfileController;
@@ -15,56 +16,43 @@ use App\Http\Controllers\KelasController;
 use App\Http\Controllers\ProdiController;
 use App\Http\Controllers\UserController;
 
-// Halaman utama
-// Route::get('/', function () {
-//     return view('auth.register');
-// });
-
 Route::get('/',[AuthController::class, 'register'])->name('register');
-// Route::get('/mahasiswa/dashboard',[DashboardController::class, 'mahasiswa'])->name('mahasiswaDashboard');
 
 Route::get('/register', [AuthController::class, 'register'])->name('auth.register');
 Route::get('/login', [AuthController::class, 'login'])->name('auth.login');
 Route::post('/register', [AuthController::class, 'store'])->name('auth.store');
 Route::post('/login', [AuthController::class, 'authenticate'])->name('auth.authenticate');
 Route::delete('/logout', [AuthController::class, 'logout'])->name('auth.logout');
-
-// Middleware untuk area login
-// Route::middleware('auth')->group(function () {
-    //     Route::get('/profile', [ProfileController::class, 'index'])->name('profile.index');
-    // });
     
-    
-Route::middleware(['auth', 'role:dosen'])->group(function () {
-        Route::get('/dosen/dashboard',[DashboardController::class, 'index'])->name('dosen.dashboard');
-        Route::get('/khs/create', [KHSController::class, 'create'])->name('dosen.khs.create');
-        Route::post('/khs/store', [KHSController::class, 'store'])->name('dosen.khs.store');
-        Route::get('/khs', [KHSController::class, 'index'])->name('dosen.khs.index');
+Route::prefix('dosen')->name('dosen.')->middleware(['auth', 'role:dosen'])->group(function () {
+        Route::get('/dashboard',[DashboardController::class, 'index'])->name('dashboard');
+        Route::get('/khs/create', [KHSController::class, 'create'])->name('khs.create');
+        Route::post('/khs/store', [KHSController::class, 'store'])->name('khs.store');
+        Route::get('/khs', [KHSController::class, 'index'])->name('khs.index');
         
         // Absensi
-        Route::get('/absensi', [AbsensiController::class, 'index'])->name('dosen.absensi.index');
-        Route::get('/absensi/create', [AbsensiController::class, 'create'])->name('dosen.absensi.create');
-        Route::post('/absensi', [AbsensiController::class, 'store'])->name('dosen.absensi.store');
-        Route::get('/absensi/{kelas_id}/edit', [AbsensiController::class, 'edit'])->name('dosen.absensi.edit');
-        Route::put('/absensi', [AbsensiController::class, 'update'])->name('dosen.absensi.update');
+        Route::get('/absensi', [AbsensiController::class, 'index'])->name('absensi.index');
+        Route::get('/absensi/create', [AbsensiController::class, 'create'])->name('absensi.create');
+        Route::post('/absensi', [AbsensiController::class, 'store'])->name('absensi.store');
+        Route::get('/absensi/{kelas_id}/edit', [AbsensiController::class, 'edit'])->name('absensi.edit');
+        Route::put('/absensi', [AbsensiController::class, 'update'])->name('absensi.update');
         
-        Route::get('/profile/dosen', [ProfileController::class, 'dosen'])->name('profile.dosen.index');
+        Route::get('/profile', [ProfileController::class, 'dosen'])->name('profile.index');
     });
     
-Route::middleware(['auth', 'role:mahasiswa'])->group(function () {
-    Route::get('/mahasiswa/dashboard',[DashboardController::class, 'index'])->name('mahasiswa.dashboard');
-    Route::get('/mahasiswa/krs', [KRSController::class, 'krsMahasiswa'])->name('krs.mahasiswa.index');
+Route::prefix('mahasiswa')->name('mahasiswa.')->middleware(['auth', 'role:mahasiswa'])->group(function () {
+    Route::get('/dashboard',[DashboardController::class, 'index'])->name('dashboard');
+    Route::get('/krs', [KRSController::class, 'index'])->name('krs.index');
     Route::get('/krs/create', [KRSController::class, 'create'])->name('krs.create');
     Route::post('/krs', [KRSController::class, 'store'])->name('krs.store');
     Route::delete('/krs/{id}', [KRSController::class, 'destroy'])->name('krs.destroy');
 
-
     Route::get('/khs/nilaiSaya', [KHSController::class, 'nilaiMahasiswa'])->name('khs.nilaiMahasiswa');
 
-    Route::get('/profile/mahasiswa', [ProfileController::class, 'mahasiswa'])->name('profile.mahasiswa.index');
+    Route::get('/profile', [ProfileController::class, 'mahasiswa'])->name('profile.index');
 });
 
-Route::middleware(['auth', 'role:admin'])->group(function () {
+Route::prefix('admin')->name('admin.')->middleware(['auth', 'role:admin'])->group(function () {
     Route::get('/admin/dashboard',[DashboardController::class, 'index'])->name('admin.dashboard');
 
     // Mahasiswa
@@ -113,13 +101,22 @@ Route::middleware(['auth', 'role:admin'])->group(function () {
     Route::delete('/kelas/{id}', [KelasController::class, 'destroy'])->name('kelas.destroy');
 
     // Prodi
-    Route::get('/prodi', [ProdiController::class, 'index'])->name('prodi.index');
-    Route::get('/prodi/create', [ProdiController::class, 'create'])->name('prodi.create');
-    Route::post('/prodi', [ProdiController::class, 'store'])->name('prodi.store');
-    Route::get('/prodi/{id}', [ProdiController::class, 'show'])->name('prodi.show');
-    Route::get('/prodi/{id}/edit', [ProdiController::class, 'edit'])->name('prodi.edit');
-    Route::put('/prodi/{id}', [ProdiController::class, 'update'])->name('prodi.update');
-    Route::delete('/prodi/{id}', [ProdiController::class, 'destroy'])->name('prodi.destroy');
+    Route::get('/prodi', [ProdiController::class, 'index'])->name('admin.prodi.index');
+    Route::get('/prodi/create', [ProdiController::class, 'create'])->name('admin.prodi.create');
+    Route::post('/prodi', [ProdiController::class, 'store'])->name('admin.prodi.store');
+    Route::get('/prodi/{id}', [ProdiController::class, 'show'])->name('admin.prodi.show');
+    Route::get('/prodi/{id}/edit', [ProdiController::class, 'edit'])->name('admin.prodi.edit');
+    Route::put('/prodi/{id}', [ProdiController::class, 'update'])->name('admin.prodi.update');
+    Route::delete('/prodi/{id}', [ProdiController::class, 'destroy'])->name('admin.prodi.destroy');
+
+    // Fakultas
+    Route::get('/fakultas', [FakultasController::class, 'index'])->name('admin.fakultas.index');
+    Route::get('/fakultas/create', [FakultasController::class, 'create'])->name('admin.fakultas.create');
+    Route::post('/fakultas', [FakultasController::class, 'store'])->name('admin.fakultas.store');
+    Route::get('/fakultas/{id}', [FakultasController::class, 'show'])->name('admin.fakultas.show');
+    Route::get('/fakultas/{id}/edit', [FakultasController::class, 'edit'])->name('admin.fakultas.edit');
+    Route::put('/fakultas/{id}', [FakultasController::class, 'update'])->name('admin.fakultas.update');
+    Route::delete('/fakultas/{id}', [FakultasController::class, 'destroy'])->name('admin.fakultas.destroy');
 
     // User
     Route::get('/user', [UserController::class, 'index'])->name('user.index');
